Security Update
--DO NOT USE IN PRODUCTION--
There is no way to delete a journalist and there is no access control regarding adding one.

This is in line with SQL injection best practices. Also an example JSON config file. I use mine for a few more API keys for other projects. Be sure to either put the JSON file in the same directory as the python file or change the filepath in the python file.